feat(enquete): refactor enquete validation (#1886)
* feat(enquete): refactor enquete validation

* feat(enquete): refactor enquete validation activite

* feat(enquete): refactor enquete validation activite

* feat(enquete): refactor enquete validation activite

* feat(enquete): refactor enquete validation prepose

* feat(enquete): refactor enquete validation prepose

* feat(enquete): refactor enquete validation prepose

* feat(enquete): refactor enquete validation prepose

+++ packages/app/src/components/Enquete/EnqueteIndividuel/mutations.js
@@ -25,7 +25,7 @@ export const CREATE_ENQUETE_INDIVIDUEL = gql`
     $anneeDebutActivite: Int
     $anneeAgrement: Int
     $cnc_annee_obtention: Int
-    $cnc_heures_formation: Int
+    $cnc_heures_formation: Float
     $niveauQualification: Int
     $niveauQualificationSecretaireSpe: Int
   ) {
